Fix to verify if the given field is belong to the given exercise before going to the edit field form

+++ app/Http/Controllers/FieldController.php
@@ -24,6 +24,7 @@ class FieldController extends Controller
     }
 
     /**
+     * show field edit form if the given field belong to the given exercise else page 404
      * @param View $view
      * @param int $id exercise id
      * @param int $fid field id
@@ -32,37 +33,45 @@ class FieldController extends Controller
      */
     public function edit(View $view, int $id, int $fid): Response
     {
-        $exercise = Exercise::find($id);
-        $field = Field::find($fid);
-        $lines = Line::all();
-        $this->displayStyle($exercise->title, 'yellow');
-
-        return $view('fields.edit', compact('lines', 'field', 'exercise'));
+        if (Field::isBelong($id, $fid)) {
+            $field = Field::find($fid);
+            $exercise = Exercise::find($id);
+            $lines = Line::all();
+            $this->displayStyle($exercise->title, 'yellow');
+            return $view('fields.edit', compact('lines', 'field', 'exercise'));
+        }
+        return $view('404');
     }
 
     /**
      * @param View $view
      * @param int $fid field id
+     * @param int $id exercise id
      * @throws ReflectionException
      */
-    public function update(View $view, int $fid): void
+    public function update(View $view, int $fid, int $id): void
     {
         Field::update($fid, $_REQUEST);
 
-        $this->redirect("/exercises");
+        $this->redirect("/exercises/$id/fields");
     }
 
     /**
+     * delete field if the exercise has building status else page 404
      * @param View $view
      * @param int $id exercise id
      * @param int $fid field id
+     * @return Response
      * @throws ReflectionException
      */
-    public function destroy(View $view, int $id, int $fid): void
+    public function destroy(View $view, int $id, int $fid): Response
     {
-        $field = Field::find($fid);
-        $field->delete();
+        if (Exercise::isEditable($id)) {
+            $field = Field::find($fid);
+            $field->delete();
 
-        $this->redirect("/exercises/$id/fields");
+            $this->redirect("/exercises/$id/fields");
+        }
+        return $view('404');
     }
 }
